requirements.txt
streamlit
PyMuPDF
python-pptx